+++ config/responsive-menus.php
@@ -7,7 +7,7 @@
  * @since   1.0.7
  * @license GPL-2.0+
  * @link    http://studiomoare.com/
- * @version 1.0.8
+ * @version 1.0.9
  */
 
 /**